+++ src/com/company/loaderPackage/Stadium.java
@@ -3,14 +3,14 @@ package com.company.loaderPackage;
 import java.util.Calendar;
 
 public class Stadium {
-    private int capacity;
+    private Integer capacity;
     private String name;
 
-    public long getCapacity() {
+    public Integer getCapacity() {
         return capacity;
     }
 
-    public void setCapacity(int capacity) {
+    public void setCapacity(Integer capacity) {
         this.capacity = capacity;
     }
